TcpSniffer: Fix not enforcing src hostname in AWS for traffic with no resolved src hostname (which is reported because might be ingress traffic)
